Unit Testing/requirements.txt
fastapi>=0.95.2,<0.96.0
httpx>=0.24.1,<0.25.0
pytest>=7.0
pytest-asyncio>=0.21
pytest-httpx>=0.22
python-dotenv>=1.0

